+++ CHANGELOG.md
@@ -17,7 +17,19 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 * Issue **#202** : Initial release of the new data retention policy functionality.
 
-## [v5.0-beta.36] - 2017-07-03
+## [v5.0-beta.39] - 2017-07-09
+
+* Issue **#341** : Fixed arg in SQL.
+
+## [v5.0-beta.38] - 2017-07-07
+
+* Issue **#340** : Fixed headless and corresponding test.
+
+## [v5.0-beta.37] - 2017-07-07
+
+* Issue **#333** : Fixed event-logging version in build.
+
+## [v5.0-beta.36] - 2017-07-06
 
 * Issue **#334** : Improved entity sorting SQL and separated generation of SQL and HQL to help avoid future issues.
 
@@ -491,7 +503,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 [Unreleased]: https://github.com/gchq/stroom/compare/v5.1-alpha.2...HEAD
 [v5.1-alpha.2]: https://github.com/gchq/stroom/compare/v5.0-alpha.1...v5.1-alpha.2
-[v5.1-alpha.1]: https://github.com/gchq/stroom/compare/v5.0-beta.36...v5.1-alpha.1
+[v5.1-alpha.1]: https://github.com/gchq/stroom/compare/v5.0-beta.39...v5.1-alpha.1
+[v5.0-beta.39]: https://github.com/gchq/stroom/compare/v5.0-beta.38...v5.0-beta.39
+[v5.0-beta.38]: https://github.com/gchq/stroom/compare/v5.0-beta.37...v5.0-beta.38
+[v5.0-beta.37]: https://github.com/gchq/stroom/compare/v5.0-beta.36...v5.0-beta.37
 [v5.0-beta.36]: https://github.com/gchq/stroom/compare/v5.0-beta.35...v5.0-beta.36
 [v5.0-beta.35]: https://github.com/gchq/stroom/compare/v5.0-beta.34...v5.0-beta.35
 [v5.0-beta.34]: https://github.com/gchq/stroom/compare/v5.0-beta.33...v5.0-beta.34